Update workitem API specs and regenerate models. - DAT-1965
The actual data returned by the workitem endpoint had become out of sync with the API spec.
The spec has been updated to reflect the data.

In the future, we might want to run tests to determine whether the data and specification still match.

The .openapi-generator-ignore file has also been moved into the api_server subfolder, since that is where the entire API lives.