Add generic 'id' field to all entities retrieved from the datastore. DAT-3142
Some entities have an id, some have name. In general that should not be a problem, since you can easily call key.id_or_name.

However, for this story, we want to be able to pass generic requests json bodies to the function, which only supports dictionaries. So, instead of key.id_or_name, we assign each entity a dictionary value of 'id'